+++ crates/biome_cli/src/execute/migrate/eslint_any_rule_to_biome.rs
@@ -1358,6 +1358,14 @@ pub(crate) fn migrate_eslint_any_rule(
             let rule = group.use_number_namespace.get_or_insert(Default::default());
             rule.set_level(rule_severity.into());
         }
+        "unicorn/throw-new-error" => {
+            if !options.include_nursery {
+                return false;
+            }
+            let group = rules.nursery.get_or_insert_with(Default::default);
+            let rule = group.use_throw_new_error.get_or_insert(Default::default());
+            rule.set_level(rule_severity.into());
+        }
         "use-isnan" => {
             let group = rules.correctness.get_or_insert_with(Default::default);
             let rule = group.use_is_nan.get_or_insert(Default::default());