+++ crates/biome_js_analyze/src/lint/nursery.rs
@@ -22,6 +22,7 @@ pub mod use_explicit_length_check;
 pub mod use_focusable_interactive;
 pub mod use_import_restrictions;
 pub mod use_sorted_classes;
+pub mod use_throw_new_error;
 
 declare_group! {
     pub Nursery {
@@ -47,6 +48,7 @@ declare_group! {
             self :: use_focusable_interactive :: UseFocusableInteractive ,
             self :: use_import_restrictions :: UseImportRestrictions ,
             self :: use_sorted_classes :: UseSortedClasses ,
+            self :: use_throw_new_error :: UseThrowNewError ,
         ]
      }
 }